Add schedule list component to routes and modify schedule helper
In this commit, a list route to the ScheduleListComponent was added to the application routes. Also, minor modifications were made in the ScheduleHelper file, primarily to handle null exception case by adding a non-null assertion operator to return statement and enhancing readability by adding a space to separate joined items.