Add ScheduleAttachSplComponent for schedule modification
This commit introduces the ScheduleAttachSplComponent, which allows the association of a Stop Point Line (SPL) with a specific schedule. The component is accessible via a new path in the application's routing module. It provides a selectable card-based view of all available SPLs, with an option to attach the selected SPL to a schedule.  A minor modification is also made to 'spl-add.component.scss'.